Create password_checker.py
This checks how many times your password has been hacked against the pwnpasswords API. We could just use the website - however, not sure how and if the passwords we type are stored and accessed by third parties. So, this is more secure when run on local machine.

Also allows you to check multiple passwords at a time and return the number of times if may have been compromised due to various data breaches from companies.